scripts/client/requirements.txt
numpy==1.16.5
Pillow==6.2.0